finished pundit

+++ app/controllers/users_controller.rb
@@ -1,5 +1,7 @@
 class UsersController < ApplicationController
     before_filter :authenticate_user!
+    # make sure you're setting authorization for each action
+    after_action :verify_authorized
     def index 
         @users = User.all
         authorize User
@@ -7,11 +9,30 @@ class UsersController < ApplicationController
     
     def show 
         @user = User.find(params[:id]) 
+        authorize @user
     end 
     
     def destroy 
-        user = User.find(parms[:id])
+        user = User.find(params[:id])
+        authorize user
         user.destroy 
-        redirect_to users_path, :notic => "User deleted"
+        redirect_to users_path, :notice => "User deleted"
     end 
+    
+    def update
+        @user = User.find(params[:id]) 
+        authorize @user
+        
+        if @user.update_attributes(secure_params)
+            redirect_to users_path, :success => 'User updated'
+        else 
+            redirect_to users_path, :alert  => 'Unable to Update User'
+        end
+    end 
+    
+    private
+        
+        def secure_params
+            params.require(:user).permit(:role)
+        end
 end
